+++ packages/storefront-next-runtime/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## v0.4.0-dev (Apr 07, 2026)
 
+- Add login preferences middleware and context to data store (@W-22051487) ([#1453](https://github.com/commerce-emu/storefront-next/pull/1453))
 - Extend `SiteProvider` to accept `site`, `language`, `locale`, `currency` props and `useSite()` to return `SiteContextValue` (@W-21787278) ([#1384](https://github.com/commerce-emu/storefront-next/pull/1384))
 - Add support to MRT Data Layer access [#1215](https://github.com/commerce-emu/storefront-next/pull/1215)
 - Add currency detection to site-context middleware (@W-21787262) ([#1342](https://github.com/commerce-emu/storefront-next/pull/1342))